+++ engine/src/systems/system_input_midi_message_out.hpp
@@ -83,8 +83,8 @@ namespace IRECS {
                 "Created MIDI device OUT {} (id: {}) on channel {} (value: {})",
                 name,
                 newDeviceId,
-                channelValue + 1,
-                channelValue
+                static_cast<int>(channelValue) + 1,
+                static_cast<int>(channelValue)
             );
             return device;
         }
